fix generators
