+++ lib/generators/carter/templates/migration.rb
